+++ src/controllers/PluginController.php
@@ -42,9 +42,13 @@ class PluginController extends Controller
 
     public function actionClearSourceCache(): Response
     {
+        $this->requirePostRequest();
         $sourceId = $this->request->getRequiredBodyParam('sourceId');
 
         Db::update('{{%video_picker_sources}}', ['cache' => null], ['id' => $sourceId]);
+
+        $source = VideoPicker::$plugin->getSources()->getSourceById((int)$sourceId);
+        $source?->clearLocalCache();
 
         Craft::$app->getSession()->setNotice(Craft::t('video-picker', 'Source cache cleared.'));